Added first version of the MCS player for the Cricket!
This version requires the Cricket! from Street Electronics to be
connected on the serial port 2 of the Apple //c. It uses the VBL
interrupt to replace the 60Hz Interrupts of the Mockingboard version so
it will run a bit slower on european models with 50Hz. This can be
partly compensated by selecting a lower TEMPO value, even tough it is
not the perfect solution yet.