MCS player for the Cricket with ACIA interrupts
The MCS-CRICKET player set the VBL as an interrupt source using the
mouse firmware. This had several limitations, since it is only
available on the //c but also because depending if the //c is a PAL or
NTSC model it would produce a 50Hz or a 60Hz interrupt.

This version uses the ACIA as an interrupt source instead making it
also compatible with other Apple II models with a Super Serial Card.
The interrupts are caused by a Cricket alarm set to 60Hz making it
region independent as well.